Update N+ license handling to reflect reporting section presence
Previously there was an assumption that `/api/9/license` will always include
the `reporting` subtree, and if it was absent, related values were populated
with zeroes [1] making the whole struct hard to rely on.

With this change, NginxLicense.Reporting will be a pointer to the original
subtree if it present in the API response, or nil otherwise.

[1] https://betterstack.com/community/guides/scaling-go/json-in-go/#2-missing-fields-fallback-to-zero-values